--- statistic_categories.json
@@ -1,269 +1,0 @@
-[
-  "main",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "main",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "main",
-  "2010",
-  "2000",
-  "main",
-  "main",
-  "main",
-  "main",
-  "main",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "race",
-  "2010",
-  "2000",
-  "2000",
-  "2000",
-  "2010",
-  "2010",
-  "race",
-  "2000",
-  "2000",
-  "2010",
-  "2010",
-  "race",
-  "2000",
-  "2000",
-  "2010",
-  "2010",
-  "race",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "national_origin",
-  "education",
-  "education",
-  "education",
-  "education",
-  "education",
-  "education",
-  "education",
-  "education",
-  "education",
-  "generation",
-  "generation",
-  "generation",
-  "generation",
-  "generation",
-  "generation",
-  "income",
-  "income",
-  "income",
-  "income",
-  "income",
-  "income",
-  "income",
-  "housing",
-  "2010",
-  "2000",
-  "housing",
-  "2010",
-  "2000",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "housing",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "transportation",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "health",
-  "climate",
-  "climate",
-  "climate",
-  "climate",
-  "climate",
-  "climate",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "industry",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "occupation",
-  "relationships",
-  "relationships",
-  "relationships",
-  "relationships",
-  "relationships",
-  "relationships",
-  "relationships",
-  "relationships",
-  "election",
-  "election",
-  "election",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "feature",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "weather",
-  "misc",
-  "misc",
-  "misc",
-  "misc",
-  "other_densities",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "other_densities",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "other_densities",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "other_densities",
-  "2010",
-  "2010",
-  "2000",
-  "2000",
-  "other_densities",
-  "other_densities"
-]
